Apply changes to more ray tests
